test mutant train and dev

+++ multi_task/mutant.py
@@ -289,7 +289,7 @@ def train_and_dev_mutant(dev_save_path_dict, dev_save_path_dataset, train_save_p
                     bag_of_CLS, type_mask, labels = dev_batch
                     bag_of_CLS = bag_of_CLS.permute(1, 0, 2)
                     type_mask = type_mask.permute(1, 0)
-                    labels = labels.permute(1, 0, 2)
+                    labels = labels.permute(1, 0, 2).type(torch.float)
 
                     with torch.no_grad():
                         outputs = model.forward(bag_of_CLS.to(device), type_mask=type_mask.to(device))
